perf
Primarily, this is to reduce the number of decisions the code needs to
make when logging something. This is done in the same way that `debug`
and some other modules do, by replacing unnecessary methods with noops
when options are changed. I added accessors for logFn and errFn so that
they always trigger a reset of the noops.